perf(core): CHECKOUT-4272 Switch to do shallow comparison by default
BREAKING CHANGE: This commit changes the way we detect changes between
the previous and the new state. Previously, we do a deep equality check
between the two by default. With this change, we do a shallow equality
check instead. This means that subscribers will get notified if any of
the immediate members of the previous state is different to the new.
This behaviour can be changed by configuring `equalityCheck` option when
creating the `DataStore` instance.